backend/requirements.txt
<<<<<<< HEAD
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
python-multipart==0.0.9
google-generativeai==0.3.2
sentence-transformers==2.5.1
faiss-cpu==1.7.4
python-docx==1.1.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
=======
fastapi
uvicorn
python-multipart
langchain
openai
>>>>>>> ea53387d94140a93c61cc8b44fec9f131457b931
